+++ src/zimscraperlib/zim/rewriting.py
@@ -89,6 +89,7 @@ def fix_target_for(
 
     # remove namespace from source ; and join it with target
     flat_target = pathlib.Path(*source.parts[1:]).parent.joinpath(target)
+
     if str(root.resolve()) == "/":
         flat_target = flat_target.relative_to(root)
     else:
@@ -199,7 +200,7 @@ def fix_urls_in_css(
 
     # split whole content on `url()` pattern to retrieve a list composed of
     # alternatively pre-pattern text and inside url() –– actual target text
-    parts = re.split(r"url\((.+)\)", content)
+    parts = re.split(r"url\((.+?)\)", content)
     for index, _ in enumerate(parts):
         if index % 2 == 0:  # skip even lines (0, 2, ..) as those are CSS code
             continue