+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * fixed URL rewriting when running from /
 * added support for link rewriting in <object> element
 * prevent from raising error if element doesn't have the attribute with url
+* use non greedy match for CSS URL links (shortest string matching `url()` format)
 
 # 1.2.0